can move around the planet in vr
