--- GE1 2022-23 Assignment Project/Assets/PlayerController.cs
@@ -1,56 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.InputSystem;
-
-public class PlayerController : MonoBehaviour
-{
-    public Rigidbody rb;
-    public GameObject camHolder;
-    public float speed, sensitivity, maxForce;
-    private Vector2 move, look;
-    private float lookRotation;
-
-    public void OnMove(InputAction.CallbackContext context)
-    {
-        move = context.ReadValue<Vector2>();
-    }
-    
-        
-    
-
-    public void OnLook(InputAction.CallbackContext context)
-    {
-        look = context.ReadValue<Vector2>();
-
-    }
-
-
-
-
-    public void FixedUpdate()
-    {
-        Vector3 currentVelocity = rb.velocity;
-        Vector3 targetVelocity = new Vector3(move.x, 0 , move.y);
-        targetVelocity *= speed;
-
-        targetVelocity = transform.TransformDirection(targetVelocity);
-
-        Vector3 velocityChange = (targetVelocity - currentVelocity);
-
-        Vector3.ClampMagnitude(velocityChange, maxForce);
-
-        rb.AddForce(velocityChange, ForceMode.VelocityChange);
-    }
-
-
-    void Start()
-    {
-        
-    }
-
-    void Update()
-    {
-        
-    }
-}
