+++ GE1 2022-23 Assignment Project/UserSettings/EditorUserSettings.asset
@@ -12,6 +12,9 @@ EditorUserSettings:
       value: 5305575e0051500f0f5b5e7516750944104f487f747876607f2b186abbb2363c
       flags: 0
     RecentlyUsedSceneGuid-2:
+      value: 550955025c07080254565d7411725b44444f1973742920617a2b1962bab33539
+      flags: 0
+    RecentlyUsedSceneGuid-3:
       value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
       flags: 0
     vcSharedLogLevel:
